backend-py/pyproject.toml
[project]
name = "backend-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "fastapi[standard]>=0.121.1",
    "psycopg2-binary>=2.9.9",
    "ultralytics>=8.2.73",
    "transformers>=4.43.3",
    "pyiqa<=0.1.14.1",
    "torch>=2.4.1",
    "Pillow>=10.4.0",
    "numpy<2.0.0",
    "timm==0.9.12",
    "python-dotenv>=1.0.1",
    "vllm>=0.6.5",
    "imagehash>=4.3.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
